Fix newlines being replaced when replacing via context menu
If using Randomize Everything via the context menu, when right-clicking at the end of a line or on a blank line, the newline would be selected and replaced.

This appears to be expected behaviour within Nova, but feels unexpected, and there is no way to circumvent this behaviour.

Thanks to @colinfoster for the bug report and suggested fix!